## Info
Replaced the list in favour of queue as it is perfectly made for this use-case. Should improve performance.

Also, this allows to queue new elements inside a queueDispatch callback.

Btw, thanks for this lib, had _exactly_ this in mind and found it then, saved me a ton of work.