@KyrietS 2025-05-11
> Also, this allows to queue new elements inside a queueDispatch callback.

Jus by looking at the code I think it's true. It would be nice to have this scenario covered in unit tests though 🙃