@timoxd7 2025-05-11
We could also test if the performance might be worse if constantly adding only one element and processing it directly. I think the performance might even be a bit worse than std::list, due to chunk allocation.

The std::deque (used by std::queue by default) allocates in chunks, not having to copy elements like a std::vector if needed and still having better cache awareness than a std::list. So it makes it a good fit for queue operations.

However, you are right, you never know until you tested it. I just added it for my code and thought it might be useful for you, but did not expect double performance (for your test in release).

I like your development style, good testing and you actually measure if something works 😄