@KyrietS 2025-05-11
According to your suggestion I modified my benchmark to make a lot of "single queues":
```cpp
	for (int i = 0; i < 2'000 * 2'0000; i++) {
		for (int k = 0; k < 1; k++) {
			dispatcher.queue(std::string { "123456789123456789123456789123456789123456789" });
		}
		dispatcher.process();
	}
```

And the results:
```
Before

Visual C++ 2022 (Debug)
81089 ms
82427 ms
85448 ms
👆😅

Visual C++ 2022 (Release)
6173 ms
5741 ms
5871 ms

GCC 13 (Debug)
677 ms
687 ms
684 ms

GCC 13 (Release)
205 ms
206 ms
202 ms

After

Visual C++ 2022 (Debug)
83678 ms
84795 ms
84459 ms

Visual C++ 2022 (Release)
1745 ms
1746 ms
1763 ms

GCC 13 (Debug)
586 ms
587 ms
588 ms

GCC 13 (Release)
106 ms
102 ms
102 ms
```

It's possible that one element is not enough to trigger a chunk deallocation so it's still faster with `std::queue`. Maybe there is a threshold somewhere, where your code becomes less performant but on a general case I like your changes and the results are great ☺️