@timoxd7 2025-05-11
@KyrietS I just added some commits (sorry for the force pushes, wanted to have a clean history but use GitHub actions for testing).

Running some local tests using clang on an arm device and -Os also had some interesting results with your test code.

Using all changes from me:
Before: ~800ms
After: ~400ms

Only using the change from std::list to std::queue, but without perfect Forward in queue(...) method:
After: ~630ms

So, the queue has less of an impact as I thought, but the forwarding (which I thought would the compiler optimize to anyway) did the bigger improvement. Maybe in real scenarios, where the nodes of a list might be even farther apart in memory than in the tests, the queue might have a bigger positive impact.